pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flet==0.25.1",
    "numpy==2.1.3",
    "openpyxl==3.1.5",
    "pandas==2.2.3",
    "psutil==6.1.0",
    "replit-object-storage>=1.0.2",
    "requests>=2.32.4",
]
